remove wlroots and move into hwnd package
