+++ core/src/cn/harryh/arkpets/ArkPets.java
@@ -10,8 +10,8 @@ import cn.harryh.arkpets.transitions.TernaryFunction;
 import cn.harryh.arkpets.transitions.TransitionFloat;
 import cn.harryh.arkpets.transitions.TransitionVector2;
 import cn.harryh.arkpets.tray.MemberTrayImpl;
-import cn.harryh.arkpets.utils.HWndCtrl;
-import cn.harryh.arkpets.utils.HWndCtrlFactory;
+import cn.harryh.arkpets.hwnd.HWndCtrl;
+import cn.harryh.arkpets.hwnd.HWndCtrlFactory;
 import cn.harryh.arkpets.utils.Logger;
 import cn.harryh.arkpets.utils.Plane;
 import com.badlogic.gdx.ApplicationAdapter;
